pyproject.toml
[project]
name = "lab-neo4j-aws"
version = "1.0.0"
description = "Neo4j and AWS Bedrock GraphRAG Workshop"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Core AWS SDK
    "boto3>=1.35.0",
    "botocore>=1.35.0",

    # Strands Agents SDK (AWS open-source agent framework)
    # See: https://strandsagents.com
    "strands-agents>=1.0.0",
    "strands-agents-tools>=1.0.0",

    # LangChain integration for Neo4j tools
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-community>=0.3.0",

    # Neo4j
    "neo4j>=5.0.0",
    "neo4j-graphrag>=1.10.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",

    # Jupyter notebooks
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

cdk = [
    "aws-cdk-lib>=2.150.0",
    "constructs>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
